Introduce MessageInterceptorAdapter for compact tests.
