+++ californium-core/src/test/java/org/eclipse/californium/core/test/ObserveTest.java
@@ -24,6 +24,7 @@
  *                                                    Introduce CountingHandler
  *                                                    use expected= annotation for
  *                                                    expected exceptions
+ *    Achim Kraus (Bosch Software Innovations GmbH) - use MessageInterceptorAdapter
  ******************************************************************************/
 package org.eclipse.californium.core.test;
 
@@ -53,7 +54,7 @@ import org.eclipse.californium.core.coap.Response;
 import org.eclipse.californium.core.network.CoapEndpoint;
 import org.eclipse.californium.core.network.EndpointManager;
 import org.eclipse.californium.core.network.config.NetworkConfig;
-import org.eclipse.californium.core.network.interceptors.MessageInterceptor;
+import org.eclipse.californium.core.network.interceptors.MessageInterceptorAdapter;
 import org.eclipse.californium.core.server.resources.CoapExchange;
 import org.eclipse.californium.rule.CoapNetworkRule;
 import org.junit.After;
@@ -390,7 +391,7 @@ public class ObserveTest {
 		uriY = String.format("coap://127.0.0.1:%d/%s", serverPort, TARGET_Y);
 	}
 
-	private class ClientMessageInterceptor implements MessageInterceptor {
+	private class ClientMessageInterceptor extends MessageInterceptorAdapter {
 
 		private int counter = 0; // counts the incoming responses
 
@@ -441,54 +442,14 @@ public class ObserveTest {
 					+ ", payload = " + response.getPayloadString());
 			response.cancel();
 		}
-
-		@Override
-		public void sendRequest(Request request) {
-		}
-
-		@Override
-		public void sendResponse(Response response) {
-		}
-
-		@Override
-		public void sendEmptyMessage(EmptyMessage message) {
-		}
-
-		@Override
-		public void receiveRequest(Request request) {
-		}
-
-		@Override
-		public void receiveEmptyMessage(EmptyMessage message) {
-		}
 	}
 
-	private class ServerMessageInterceptor implements MessageInterceptor {
+	private class ServerMessageInterceptor extends MessageInterceptorAdapter {
 
 		private final AtomicInteger resetCounter;
 
 		public ServerMessageInterceptor(AtomicInteger resetCounter) {
 			this.resetCounter = resetCounter;
-		}
-
-		@Override
-		public void receiveResponse(Response response) {
-		}
-
-		@Override
-		public void sendRequest(Request request) {
-		}
-
-		@Override
-		public void sendResponse(Response response) {
-		}
-
-		@Override
-		public void sendEmptyMessage(EmptyMessage message) {
-		}
-
-		@Override
-		public void receiveRequest(Request request) {
 		}
 
 		@Override
